Rename 중복 파일명 index.jsx로 변경
