+++ fe/src/components/IssueTable/TableToolBar/IssueListFilters/TableFilterButtons/index.jsx
@@ -13,7 +13,7 @@ import {
   MainPageContext,
   MainPageDispatchContext,
 } from '../../../../../context/MainPage/MainPageContext';
-import DropDown from '../../../../DropDown/DropDown';
+import DropDown from '../../../../DropDown';
 
 // TODO(덴): DropDown Config 변환해주는 함수 유틸로 빼보기.
 const convertOptionItems = (
