+++ fe/src/components/common/Button/index.js
@@ -1,3 +1,3 @@
-export * from './ContainerButton';
-export * from './GhostButton';
-export * from './OutLineButton';
+export { default as ContainerButton } from './ContainerButton';
+export { default as GhostButton } from './GhostButton';
+export { default as OutlineButton } from './OutlineButton';
