+++ fe/src/components/common/PageLayout/index.jsx
@@ -1,6 +1,6 @@
 import { styled } from 'styled-components';
 import { Outlet } from 'react-router-dom';
-import { Header } from '../Header/Header';
+import { Header } from '../Header';
 
 export const PageLayout = () => {
   return (
