+++ fe/src/components/common/index.js
@@ -1,7 +1,7 @@
-export * from './Header/Header';
-export * from './PageLayout/PageLayout';
-export * from './UserIcon/UserIcon';
-export * from './Label/Label';
-export * from './ChevronDown/ChevronDown';
-export * from './Spinner/Spinner';
+export * from './Header';
+export * from './PageLayout';
+export * from './UserIcon';
+export * from './Label';
+export * from './ChevronDown';
+export * from './Spinner';
 export * from './Button';
